refactor tests structure, one more time
